modified TeamService and Dto after I removed team_rating field

+++ teamUp/src/main/java/com/teamup/teamUp/model/dto/rating/team/TeamRatingDto.java
@@ -4,6 +4,6 @@ public record TeamRatingDto(
         int attack,
         int midfield,
         int defense,
-        double overall
+        int overall
 ) {
 }
